refactor: dodanie usuwania pliku z minio
