+++ src/main/java/org/aszjch/demoapp/domain/article/ArticleConfig.java
@@ -1,9 +1,8 @@
-package org.aszjch.demoapp.infrastructure.config;
+package org.aszjch.demoapp.domain.article;
 
-import org.aszjch.demoapp.domain.article.ArticleService;
-import org.aszjch.demoapp.domain.article.ArticleServiceImpl;
 import org.aszjch.demoapp.domain.article.port.ArticlePublisher;
 import org.aszjch.demoapp.domain.article.port.ArticleRepository;
+import org.aszjch.demoapp.domain.articlefile.ArticleFileService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -12,8 +11,9 @@ class ArticleConfig {
 
     @Bean
     ArticleService articleService(final ArticleRepository articleRepository,
-                                  final ArticlePublisher articlePublisher) {
-        return new ArticleServiceImpl(articleRepository, articlePublisher);
+                                  final ArticlePublisher articlePublisher,
+                                  final ArticleFileService articleFileService) {
+        return new ArticleServiceImpl(articleRepository, articlePublisher, articleFileService);
     }
 
 }
