+++ src/main/java/org/aszjch/demoapp/domain/articlefile/ArticleFileConfig.java
@@ -1,8 +1,6 @@
-package org.aszjch.demoapp.infrastructure.config;
+package org.aszjch.demoapp.domain.articlefile;
 
 import org.aszjch.demoapp.domain.article.ArticleService;
-import org.aszjch.demoapp.domain.articlefile.ArticleFileService;
-import org.aszjch.demoapp.domain.articlefile.ArticleFileServiceImpl;
 import org.aszjch.demoapp.domain.articlefile.port.ArticleFileAssignmentService;
 import org.aszjch.demoapp.domain.articlefile.port.ArticleFileStorageService;
 import org.springframework.context.annotation.Bean;
