+++ src/main/java/org/aszjch/demoapp/domain/articlefile/ArticleFileService.java
@@ -3,4 +3,6 @@ package org.aszjch.demoapp.domain.articlefile;
 public interface ArticleFileService {
 
     void upload(ArticleFile articleFile);
+
+    void delete(ArticleFile articleFile);
 }
