+++ src/main/java/org/aszjch/demoapp/domain/articlefile/port/ArticleFileStorageService.java
@@ -5,4 +5,6 @@ import org.aszjch.demoapp.domain.articlefile.ArticleFile;
 public interface ArticleFileStorageService {
 
     void store(ArticleFile articleFile);
+
+    void delete(ArticleFile articleFile);
 }
