+++ src/main/java/org/aszjch/demoapp/infrastructure/storage/ExternalStorageClient.java
@@ -1,9 +1,12 @@
 package org.aszjch.demoapp.infrastructure.storage;
 
 import io.minio.ObjectWriteResponse;
+import org.aszjch.demoapp.domain.articlefile.ArticleFile;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface ExternalStorageClient {
 
     ObjectWriteResponse upload(MultipartFile file);
+
+    void delete(ArticleFile articleFile);
 }
